Use all node children when generating post listing
Instead of finding a node with a parent of a specific type (heading,
paragraph) and using its literal, find the node type we want, and
collect all its children's literals.  A paragraph is not a single node,
so this will make the summary (first paragraph) extraction more robust.

This process drops all formatting, however.